+++ DVehiculos.php
@@ -1,4 +1,14 @@
 <?php
+session_start();
+
+// Verificar si el usuario está autenticado
+if (!isset($_SESSION['authenticated']) || $_SESSION['authenticated'] !== true) {
+    // Redirigir al usuario a la página de inicio de sesión
+    header('Location: FAcceso.php');
+    exit;
+}
+?>
+<?php
     include("Controlador.php");
     $NIV = $_GET['NIV'];
     $SQL = "DELETE FROM Vehiculos WHERE NIV='$NIV'";
